Fix for BL-touch users.
 RiskyBirdness 16th April 2019 at 00:58 GMT
Discord

PUBLIC SERVICE ANNOUNCEMENT FOR BLTOUCH USERS.. When you power up your printer the BLTouch runs an internal test routine and deploys the probe twice. 
If the nozzle height is close to the bed at the time the routine is run the probe will not fully deploy and BLTouch will enter an alarm state(flashing). 
In this state it will be non functional but the Duet has no way of knowing this. If you perform a Z home in this state you will get a nozzle crash, 
damage your BLTouch probe, and damage your print surface. 
The solution is to always clear your BLTouch alarm at the top of your homeZ.g and homeall.g scripts using M280 P3 S160.

Not sure whether to uncomment it and make it "standard" or not. Will have to test to see if it affects other probes such as IR.